lift/index.html

<html>
  <head>
    <link href="../tester/testharness.css" rel="stylesheet">
    <script src="../tester/testharness.js"></script>
  </head>
  <body>
    <style>
      body{
        margin: 0;
      }

      .wrapper{
        max-width: 1024px;
        margin: 0 auto;
        position: relative;
        height: 100%;
      }

      .lift-cabine{
        border: 4px solid black;
        width: 150px;
        height: 200px;
        position: absolute;
        left: 50%;
        transform: translateX(-50%);
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
        bottom: 0;
      }

      .lift-cabine.basement{
        bottom: -190px;
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
      }

      .lift-cabine.first{
        bottom: 0;
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
      }


      .lift-cabine.second{
        bottom: 50%;
        transition: bottom 2s ease-in-out;
        -webkit-transition: bottom 2s ease-in-out;
      }


      .lift-cabine.third{
        bottom: calc(100% - 210px);
      }

      .buttons{
        width: 50px;
        position: absolute;
        right: 0;
        top: 50%;
        transform: translateY(-50%);
      }

      .buttons button{
        width: 50px;
        height: 50px;
        margin-bottom: 10px;
      }
    </style>
    <div class='wrapper'>
      <div class='lift-cabine second'></div>
      <div class='buttons'>
        <button id='3'>3</button>
        <button id='2'>2</button>
        <button id='1'>1</button>
        <button id='-1'>-1</button>
      </div>
    </div>


<!--testing-->
    <script>
      var buttons = document.querySelectorAll('button');

      var test1 = async_test('testing querySelectorAll for floors-buttons');

      window.addEventListener('load', test1.step_func(function(e){
        var amount = 4;
        assert_equals(buttons.length, amount);
        this.done();
      }))
    </script>
  </body>
</html>
